+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 0.1.1 - `--reset` cli arg, better messages
 
-- Add custom responses for cache misses while network access is disabled
-- Add custom responses for unmocked requests while network access is disabled
-- Add `--reset`/`-r` reset flag to restart jambox
+- feat: custom responses for cache misses while network access is disabled
+- feat: custom responses for unmocked requests while network access is disabled
+- feat: `--reset`/`-r` reset flag to restart jambox
+- fix: allow `localhost` requests even if network access is disabled #42
 
 ## 0.1.0 - Refactor